무한참조 버그픽스 + 리뷰좋아요 구현중

+++ src/main/java/com/example/ajouparking/controller/LikesController.java
@@ -22,7 +22,7 @@ public class LikesController {
     private final LikesService likesService;
     private final ReviewRepository reviewRepository;
     @PostMapping("/{toReviewId}")
-    public ResponseEntity<?> like(@AuthenticationPrincipal CustomUserDetails customUserDetails, @PathVariable int toReviewId){
+    public ResponseEntity<?> like(@AuthenticationPrincipal CustomUserDetails customUserDetails, @PathVariable Long toReviewId){
         User user = customUserDetails.getUser();
         likesService.putLike(user.getId(),toReviewId);
 
@@ -31,9 +31,10 @@ public class LikesController {
     }
 
     @DeleteMapping("/{toReviewId}")
-    public ResponseEntity<?> unlike(@AuthenticationPrincipal CustomUserDetails customUserDetails, @PathVariable int toReviewId){
+    public ResponseEntity<?> unlike(@AuthenticationPrincipal CustomUserDetails customUserDetails, @PathVariable Long toReviewId){
         User user = customUserDetails.getUser();
         likesService.deleteLike(user.getId(), toReviewId);
         return new ResponseEntity<>(new CommonResponseDto<>("삭제성공",null),HttpStatus.NO_CONTENT);
     }
+
 }
