+++ src/main/java/com/example/ajouparking/repository/LikesRepository.java
@@ -9,9 +9,9 @@ public interface LikesRepository extends JpaRepository<Likes,Long> {
 
     @Modifying
     @Query(value = "insert into likes(from_user_id,to_review_id) values(:fromUserId,:toReviewId)",nativeQuery = true)
-    void like(int fromUserId,int toReviewId);
+    void like(int fromUserId,Long toReviewId);
 
     @Modifying
     @Query(value = "delete from likes where from_user_id = :fromUserId and to_review_id = :toReviewId",nativeQuery=true)
-    void unlike(int fromUserId, int toReviewId);
+    void unlike(int fromUserId, Long toReviewId);
 }
